Create Encryptor class
