+++ lib/enigma.rb
@@ -1,16 +1,17 @@
 require 'date'
+require './lib/generatable'
 
 class Enigma
+  include Generatable
   attr_reader :character_set
 
   def initialize()
-    @character_set = ("a".."z").to_a << " "
+    @encryptor = Encryptor.new()
   end
 
   def encrypt(message, key = generate_random_key_number(), date = Date.today.strftime("%d%m%y"))
     {
-      encryption: get_encrypted_string(message.downcase, generate_keys(key),
-                  generate_offsets(date)),
+      encryption: @encryptor.get_encrypted_string(message.downcase, key, date),
       key: key,
       date: date
     }
@@ -23,22 +24,6 @@ class Enigma
     key: key,
     date: date
     }
-  end
-
-  def get_encrypted_string(message, keys, offsets)
-    shifts = generate_shifts(keys, offsets)
-    encrypted_array = []
-    message.split('').each_slice(4) do |four_chars|
-      encrypted_array << find_encrypted_letters(four_chars, shifts)
-    end
-    encrypted_array.flatten.join
-  end
-
-  def find_encrypted_letters(four_char_arr, shifts)
-    four_char_arr.zip(shifts.values).map do |char, shift_value|
-      next char if !@character_set.include?(char)
-      @character_set.rotate(@character_set.index(char) + shift_value).first
-    end
   end
 
   def get_decrypted_string(message, keys, offsets)
